Combine Home files
